requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sock==0.5.2
openai-whisper==20231117
numpy==1.21.2
torch==2.0.1
soundfile==0.10.3.post1 